+++ Flow/src/utopia/flow/collection/immutable/caching/cache/Cache.scala
@@ -1,8 +1,8 @@
 package utopia.flow.collection.immutable.caching.cache
 
-import utopia.flow.collection.template.CacheLike
+import utopia.flow.collection.template.MapAccess
 
-import scala.collection.immutable.HashMap
+import scala.collection.mutable
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.{Duration, FiniteDuration}
 
@@ -15,7 +15,7 @@ object Cache
 	  * @tparam Value The type of cached result
 	  * @return A new cache
 	  */
-	def apply[Key, Value](request: Key => Value) = new Cache[Key, Value](request)
+	def apply[Key, Value](request: Key => Value): Cache[Key, Value] = new _Cache[Key, Value](request)
 	
 	/**
 	  * Creates a weakly referencing cache
@@ -75,35 +75,68 @@ object Cache
 	                                  (request: K => V)
 	                                  (implicit exc: ExecutionContext) =
 		ReleasingCache.after(threshold)(request)
+		
+	
+	// NESTED   ---------------------------
+	
+	private class _Cache[Key, Value](request: Key => Value) extends Cache[Key, Value]
+	{
+		// ATTRIBUTES	------------------
+		
+		private val cachedItems: mutable.Map[Key, Value] = mutable.Map()
+		
+		
+		// IMPLEMENTED	------------------
+		
+		override def cachedValues = cachedItems.values
+		
+		override def apply(key: Key) = cachedItems.getOrElseUpdate(key, request(key))
+		override def cached(key: Key) = cachedItems.get(key)
+	}
 }
 
 /**
-  * This is a simple implementation of the CacheLike trait
+  * Caches keep their values after requesting them
   * @author Mikko Hilpinen
   * @since 10.6.2019, v1.5+
   */
-class Cache[Key, Value](private val request: Key => Value) extends CacheLike[Key, Value]
+trait Cache[-Key, +Value] extends MapAccess[Key, Value]
 {
-	// ATTRIBUTES	------------------
+	// ABSTRACT ---------------------------
 	
-	private var cachedItems: Map[Key, Value] = HashMap()
+	/**
+	  * @return All the items currently stored in this cache
+	  */
+	def cachedValues: Iterable[Value]
+	
+	/**
+	  * The currently cached value for specified key
+	  * @param key A key
+	  * @return The currently cached value for the key. None if no data is cached
+	  */
+	def cached(key: Key): Option[Value]
 	
 	
-	// IMPLEMENTED	------------------
+	// OTHER    --------------------------
 	
-	override def cachedValues = cachedItems.values
+	/**
+	  * @param key A key
+	  * @return Whether there is currently cached data for the key
+	  */
+	def isValueCached(key: Key) = cached(key).isDefined
 	
-	override def apply(key: Key) =
-	{
-		if (cachedItems.contains(key))
-			cachedItems(key)
-		else
-		{
-			val value = request(key)
-			cachedItems += key -> value
-			value
-		}
-	}
-	
-	override def cached(key: Key) = cachedItems.get(key)
+	/**
+	  * @param f A mapping function applied for values of this cache
+	  * @tparam O  Type of values accepted by the mapping function (i.e. Value of this cache or higher type)
+	  * @tparam V2 Type of values returned by mapping function
+	  * @return A view of this cache that maps returned values (but doesn't cache mapped values)
+	  */
+	def mapValuesView[O >: Value, V2](f: O => V2) = MappingCacheView[Key, O, V2](this)(f)
+	/**
+	  * @param f A mapping function applied for key inputs to this cache, producing a key applicable to this cache
+	  * @tparam K2 Type of keys accepted by the resulting cache (map input)
+	  * @tparam OK Type of keys produced by the mapping function (map output, must be a sub-type of this cache's key)
+	  * @return A copy of this cache that maps the incoming keys before proposing them to this cache
+	  */
+	def mapKeys[K2, OK <: Key](f: K2 => OK) = KeyMappingCache[K2, OK, Value](this)(f)
 }