+++ Flow/src/utopia/flow/collection/immutable/caching/cache/KeyMappingCache.scala
@@ -1,7 +1,5 @@
 package utopia.flow.collection.immutable.caching.cache
 
-import utopia.flow.collection.template.CacheLike
-
 object KeyMappingCache
 {
 	/**
@@ -12,7 +10,7 @@ object KeyMappingCache
 	  * @tparam V Type of values returned by these caches
 	  * @return A new key-mapping cache
 	  */
-	def apply[K, KO, V](wrapped: CacheLike[KO, V])(f: K => KO) = new KeyMappingCache[K, KO, V](wrapped)(f)
+	def apply[K, KO, V](wrapped: Cache[KO, V])(f: K => KO) = new KeyMappingCache[K, KO, V](wrapped)(f)
 }
 
 /**
@@ -20,11 +18,10 @@ object KeyMappingCache
   * @author Mikko Hilpinen
   * @since 26.7.2022, v1.16
   */
-class KeyMappingCache[-K, KO, +V](wrapped: CacheLike[KO, V])(f: K => KO) extends CacheLike[K, V]
+class KeyMappingCache[-K, KO, +V](wrapped: Cache[KO, V])(f: K => KO) extends Cache[K, V]
 {
 	override def cachedValues = wrapped.cachedValues
 	
 	override def cached(key: K) = wrapped.cached(f(key))
-	
 	override def apply(key: K) = wrapped(f(key))
 }